+++ internal-test-support/src/main/java/com/wealthfront/magellan/internal/test/AttachingActivity.kt
@@ -1,0 +1,17 @@
+package com.wealthfront.magellan.internal.test
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import com.wealthfront.magellan.core.Navigable
+import com.wealthfront.magellan.lifecycle.setContentScreen
+
+public class AttachingActivity : ComponentActivity() {
+  public companion object {
+    public var step: Navigable? = null
+  }
+
+  override fun onCreate(savedInstanceState: Bundle?) {
+    super.onCreate(savedInstanceState)
+    setContentScreen(step!!)
+  }
+}