+++ internal-test-support/src/main/java/com/wealthfront/magellan/internal/test/DummyStep.kt
@@ -1,0 +1,6 @@
+package com.wealthfront.magellan.internal.test
+
+import com.wealthfront.magellan.core.Step
+import com.wealthfront.magellan.internal.test.databinding.MagellanDummyLayoutBinding
+
+public open class DummyStep : Step<MagellanDummyLayoutBinding>(MagellanDummyLayoutBinding::inflate)