+++ internal-test-support/src/main/java/com/wealthfront/magellan/test/DummyStep.kt
@@ -1,6 +1,0 @@
-package com.wealthfront.magellan.test
-
-import com.wealthfront.magellan.core.Step
-import com.wealthfront.magellan.test.databinding.MagellanDummyLayoutBinding
-
-public open class DummyStep : Step<MagellanDummyLayoutBinding>(MagellanDummyLayoutBinding::inflate)