+++ magellan-legacy/src/test/java/com/wealthfront/magellan/LegacyViewComponentTest.java
@@ -4,7 +4,7 @@ import android.app.Activity;
 import android.content.Context;
 import android.util.SparseArray;
 
-import com.wealthfront.magellan.test.DummyScreen;
+import com.wealthfront.magellan.internal.test.DummyScreen;
 
 import org.junit.Before;
 import org.junit.Test;