+++ magellan-library/src/main/java/com/wealthfront/magellan/navigation/LinearNavigator.kt
@@ -1,40 +1,19 @@
 package com.wealthfront.magellan.navigation
 
 import com.wealthfront.magellan.Direction
-import com.wealthfront.magellan.OpenForMocking
-import com.wealthfront.magellan.ScreenContainer
 import com.wealthfront.magellan.core.Navigable
-import com.wealthfront.magellan.lifecycle.LifecycleAwareComponent
-import com.wealthfront.magellan.lifecycle.attachFieldToLifecycle
+import com.wealthfront.magellan.lifecycle.LifecycleAware
 import com.wealthfront.magellan.transitions.MagellanTransition
 import java.util.Deque
 
-@OpenForMocking
-public class LinearNavigator internal constructor(
-  container: () -> ScreenContainer
-) : Navigator, LifecycleAwareComponent() {
+public interface LinearNavigator : Navigator, LifecycleAware {
 
-  private val delegate by attachFieldToLifecycle(NavigationDelegate(container))
+  public fun goTo(navigable: Navigable, overrideMagellanTransition: MagellanTransition? = null)
 
-  override val backStack: List<NavigationEvent>
-    get() = delegate.backStack.toList()
-
-  public fun goTo(navigable: Navigable, overrideMagellanTransition: MagellanTransition? = null) {
-    delegate.goTo(navigable, overrideMagellanTransition)
-  }
-
-  public fun replace(navigable: Navigable, overrideMagellanTransition: MagellanTransition? = null) {
-    delegate.replace(navigable, overrideMagellanTransition)
-  }
+  public fun replace(navigable: Navigable, overrideMagellanTransition: MagellanTransition? = null)
 
   public fun navigate(
     direction: Direction,
     backStackOperation: (Deque<NavigationEvent>) -> MagellanTransition
-  ) {
-    delegate.navigate(direction, backStackOperation)
-  }
-
-  public override fun goBack(): Boolean {
-    return delegate.goBack()
-  }
+  )
 }