+++ magellan-library/src/main/java/com/wealthfront/magellan/navigation/NavigationDelegate.kt
@@ -114,25 +114,28 @@ public class NavigationDelegate(
     oldBackStack: List<NavigableCompat>,
     newBackStack: List<NavigableCompat>
   ) {
-    val oldNavigables = oldBackStack.toSet()
-    val newNavigables = newBackStack.toSet()
-    if (newNavigables.size != newBackStack.size) {
-      val numExtraNavigables = newNavigables.size - newBackStack.size
+    val oldNavigableSet = oldBackStack.toSet()
+    val newNavigableSet = newBackStack.toSet()
+    if (newNavigableSet.size != newBackStack.size) {
+      val numExtraNavigables = newBackStack.size - newNavigableSet.size
       throw IllegalStateException(
         "Cannot have multiple of the same Navigable in the backstack. " +
-          "Have $numExtraNavigables extra Navigables in: $newNavigables"
+          "Have $numExtraNavigables extra Navigables in: ${newBackStack.toReadableString()}"
       )
     }
 
-    (oldNavigables - newNavigables).forEach { oldNavigable ->
+    (oldNavigableSet - newNavigableSet).forEach { oldNavigable ->
       lifecycleRegistry.removeFromLifecycle(oldNavigable)
     }
 
-    (newNavigables - oldNavigables).forEach { newNavigable ->
+    (newNavigableSet - oldNavigableSet).forEach { newNavigable ->
       currentNavigableSetup?.invoke(newNavigable)
       lifecycleRegistry.attachToLifecycleWithMaxState(newNavigable, CREATED)
     }
   }
+
+  private fun List<NavigableCompat>.toReadableString() =
+    joinToString(prefix = "[", postfix = "]") { it::class.java.simpleName }
 
   private fun animateAndRemove(
     from: View?,
@@ -148,7 +151,7 @@ public class NavigationDelegate(
         NoAnimationTransition()
       }
       transition.animate(from, to, direction) {
-        if (context != null) {
+        if (context != null && containerView != null) {
           containerView!!.removeView(from)
           currentNavigable!!.transitionFinished()
           containerView!!.setInterceptTouchEvents(false)