+++ magellan-library/src/main/java/com/wealthfront/magellan/transitions/DefaultTransition.kt
@@ -6,8 +6,16 @@ import android.animation.AnimatorSet
 import android.animation.ObjectAnimator
 import android.util.Property
 import android.view.View
+import androidx.interpolator.view.animation.FastOutSlowInInterpolator
 import com.wealthfront.magellan.Direction
+import com.wealthfront.magellan.navigation.NavigationEvent
 
+/**
+ * The default transition for all [NavigationEvent]s where another [MagellanTransition] isn't
+ * defined. Performs a right-to-left slide on entrance and a left-to-right slide on exit. Uses a
+ * [FastOutSlowInInterpolator] for both per
+ * [the Material Design guidelines](https://material.io/design/motion/speed.html#easing).
+ */
 public class DefaultTransition : MagellanTransition {
 
   override fun animate(
@@ -39,6 +47,7 @@ public class DefaultTransition : MagellanTransition {
       set.play(ObjectAnimator.ofFloat(from, axis, 0f, fromTranslation.toFloat()))
     }
     set.play(ObjectAnimator.ofFloat(to, axis, toTranslation.toFloat(), 0f))
+    set.interpolator = FastOutSlowInInterpolator()
     return set
   }
 }