+++ magellan-library/src/test/java/com/wealthfront/magellan/navigation/NavigationTraverserTest.kt
@@ -5,9 +5,9 @@ import android.content.Context
 import com.google.common.truth.Truth.assertThat
 import com.wealthfront.magellan.core.Journey
 import com.wealthfront.magellan.core.Step
+import com.wealthfront.magellan.internal.test.databinding.MagellanDummyLayoutBinding
 import com.wealthfront.magellan.lifecycle.LifecycleState.Created
 import com.wealthfront.magellan.lifecycle.transitionToState
-import com.wealthfront.magellan.test.databinding.MagellanDummyLayoutBinding
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith